pyproject.toml
[tool.poetry]
name = "sentinel"
version = "0.5.1-alpha"
license = "GPL-3.0-or-later"
description = ""
authors = ["Anmol Brar <brarsanmol@protonmail.com>"]
readme = "README.md"
homepage = "https://sentinel.anmolbrar.ca"
repository = "https://github.com/brarsanmol/sentinel"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = {git = "https://github.com/Rapptz/discord.py.git", rev = "master"}
sendgrid = "^6.9.7"
pugsql = "^0.2.4"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
prospector = {extras = ["with_pyroma"], version = "^1.7.7"}
pre-commit = "^2.17.0"
isort = "^5.10.1"
commitizen = "^2.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
